run lab on all platforms in containers
